style: Refactor test naming
