+++ tests/LSP4-DigitalAsset-Metadata/LSP4-Compatibility.spec.ts
