+++ tests/LSP7-DigitalAsset/LSP7-DigitalAsset.spec.ts
@@ -6,7 +6,7 @@ import {
   shouldBehaveLikeLSP7,
   shouldInitializeLikeLSP7,
   LSP7TestContext,
-} from "./LSP7.behaviour";
+} from "./LSP7-DigitalAsset.behaviour";
 
 describe("LSP7", () => {
   describe("when using LSP7 contract with constructor", () => {
