+++ tests/LSP7-DigitalAsset/extensions/LSP7-CappedSupply.behaviour.ts
