+++ tests/LSP7-DigitalAsset/extensions/LSP7-CompatabilityForERC20.spec.ts
