+++ tests/LSP8-IdentifiableDigitalAsset/LSP8-IdentifiableDigitalAsset.behaviour.ts
