+++ tests/LSP8-IdentifiableDigitalAsset/extensions/LSP8-CappedSupply.behaviour.ts
