+++ tests/LSP8-IdentifiableDigitalAsset/extensions/LSP8-CappedSupply.spec.ts
@@ -8,8 +8,8 @@ import {
   shouldBehaveLikeLSP8CappedSupply,
   LSP8CappedSupplyTestContext,
   getNamedAccounts,
-} from "./LSP8CappedSupply.behaviour";
-import { shouldInitializeLikeLSP8 } from "../LSP8.behaviour";
+} from "./LSP8-CappedSupply.behaviour";
+import { shouldInitializeLikeLSP8 } from "../LSP8-IdentifiableDigitalAsset.behaviour";
 
 describe("LSP8CappedSupply", () => {
   describe("when using LSP8CappedSupply with constructor", () => {
