+++ tests/LSP8-IdentifiableDigitalAsset/extensions/LSP8-CompatibilityForERC721.spec.ts
